% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.r
\name{life}
\alias{life}
\title{Play Conway's Game of Life}
\usage{
life(state, ..., gen_max = NULL, sleep = NULL)
}
\arguments{
\item{state}{A matrix representing the initial state.}

\item{...}{Graphics options passed down to \code{image}.}

\item{gen_max}{The maximum number of generations to compute. Or \code{NULL} to
run indefinitely.}

\item{sleep}{The sleep interval between generations, in seconds. Or \code{NULL}
for no sleep interval.}
}
\description{
Given a starting state, this function will display each generation of
Conway's Game of Life on the active graphics device. The state is updated
according to the rules in the Rules section.
}
\section{Rules}{

\enumerate{
\item Any live cell with two or three neighbors survives.
\item Any dead cell with three live neighbors becomes a live cell.
\item All other live cells die in the next generation. Similarly,
all other dead cells stay dead.
}
}

