% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.r
\name{state_init}
\alias{state_init}
\title{Initialize a matrix to play Conway's Game of Life}
\usage{
state_init(r, c, p)
}
\arguments{
\item{r, c}{The number of rows and columns.}

\item{p}{The proportion of "living" cells.}
}
\description{
Initialize a matrix to play Conway's Game of Life
}
